requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
better-profanity==0.7.0
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
yarl==1.20.1
